remove backup file
