+++ src/drt/src/gr/FlexGR_util.cpp
@@ -1,6 +1,6 @@
 /* Authors: Zhiang Wang */
 /*
- * Copyright (c) 2024, The Regents of the University of California
+ * Copyright (c) 2025, The Regents of the University of California
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
