+++ src/drt/src/gr/FlexGR_util.h
@@ -1,6 +1,6 @@
 /* Authors: Zhiang Wang */
 /*
- * Copyright (c) 2024, The Regents of the University of California
+ * Copyright (c) 2025, The Regents of the University of California
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,13 +30,23 @@
 
 
 
+#include "odb/db.h"
+#include "db/grObj/grShape.h"
+#include "db/grObj/grVia.h"
 #include "db/grObj/grPin.h"
+#include "db/infra/frTime.h"
+#include "db/obj/frGuide.h"
 #include "dr/FlexMazeTypes.h"
+#include "utl/exception.h"
+#include "stt/SteinerTreeBuilder.h"
+#include "gr/FlexGRWavefront.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "gr/FlexGRWavefront.h"
 
 #include <iostream>
+#include <queue>
+#include <thread>
+#include <tuple>
 #include <sys/resource.h>
 #include <cuda_runtime.h>
 #include <cuda.h>
@@ -47,254 +57,155 @@
 
 namespace drt {
 
+// ------------------------------------------------------------------------------
+// Hyperparameters for the FlexGR
+// ------------------------------------------------------------------------------
+constexpr int GRGRIDGRAPHHISTCOSTSIZE = 8;
+constexpr int GRSUPPLYSIZE = 8;
+constexpr int GRDEMANDSIZE = 16;
+constexpr int GRFRACSIZE = 1;
+
+// -------------------------------------------------------------------------------
+// Some basic structures used in the FlexGR
+// -------------------------------------------------------------------------------
+
+
+// In the ripup and reroute stage, we need to keep tracking gcells with overflow
+// and the nets passing through the grid
+struct GridStruct {
+  int xIdx = -1;
+  int yIdx = -1;
+  int zIdx = 0; // Default is 2D grid
+
+  bool isOverflowH = false;
+  bool isOverflowV = false;
+
+  std::set<frNet*> nets; // All the nets passing through the grid
+  
+  GridStruct() = default;
+  
+  GridStruct(int _xIdx, int _yIdx, int _zIdx) 
+    : xIdx(_xIdx), yIdx(_yIdx), zIdx(_zIdx) { }
+    
+};
+
+
+
+// This is used for the net-level parallelization
+struct NetStruct {
+  frNet* net = nullptr;
+  int netId = -1;
+  int nodeCntStart = -1; // starting node index in the net
+  int numNodes = -1;
+
+  std::vector<int> points; // Store all the steiner points in the net
+  std::vector<std::pair<int, int> > vSegments;
+  std::vector<std::pair<int, int> > hSegments;
+
+  NetStruct() = default;
+
+  NetStruct(frNet* _net, int _netId, int _nodeCntStart, int _numNodes) 
+    : net(_net), netId(_netId), nodeCntStart(_nodeCntStart), numNodes(_numNodes) { }
+ 
+};
+
+
+
+
+struct IntPair {
+  int start = -1;
+  int end = -1;
+
+  IntPair() = default;
+
+  IntPair(int _start, int _end) : start(_start), end(_end) { }
+
+  __host__ __device__
+  int x() const { return start; }
+  
+  __host__ __device__
+  int y() const { return end; }
+
+  __host__ __device__
+  int start() const { return start; }
+
+  __host__ __device__
+  int end() const { return end; }
+
+  __host__ __device__
+  void set(int _start, int _end) {
+    start = _start;
+    end = _end;
+  }
+};
+
+
 void printPeakMemoryUsage();
 
 
-__host__ __device__
-bool addEdge(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
+// -----------------------------------------------------------------------------
+// CUDA Related Device functions
+// -----------------------------------------------------------------------------
 
-__host__ __device__
-bool removeEdge(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-__host__ __device__
-bool setBlock(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-__host__ __device__
-bool resetBlock(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-__host__ __device__
-void setHistoryCost(uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, unsigned histCostIn);
-
-__host__ __device__
-void addHistoryCost(uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, unsigned in = 1);
-
-__host__ __device__
-void decayHistoryCost(uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z);
-
-__host__ __device__
-void decayHistoryCost(uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, double d);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void setSupply(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned supplyIn);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void setDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned demandIn);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void setRawDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned rawDemandIn);
-
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void addDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned delta = 1);
-
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void addRawDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned delta = 1);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void subDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned delta = 1);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-void subRawDemand(uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, unsigned delta = 1);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-unsigned getSupply(const uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir, bool isRaw = false);
-
-
-__host__ __device__
-unsigned getRawSupply(const uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-unsigned getDemand(const uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-// E == H; N == V; currently no U / D
-__host__ __device__
-unsigned getRawDemand(const uint64_t* bits,
-  const bool* zDir, int xDim, int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-__host__ __device__
-bool hasEdge(const uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-__host__ __device__
-bool hasBlock(const uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-__host__ __device__
-bool hasHistoryCost(const uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z);
-
-
-__host__ __device__
-bool hasCongCost(const uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-__host__ __device__
-unsigned getHistoryCost(const uint64_t* bits, 
-  const bool* zDir, int xDim, int yDim, int zDim, 
-  frMIdx x, frMIdx y, frMIdx z);
-
-__host__ __device__
-frCoord getEdgeLength(
-  const frCoord* xCoords, const frCoord* yCoords, const frCoord* zHeights,
-  const bool* zDirs, int xDim , int yDim, int zDim,
-  frMIdx x, frMIdx y, frMIdx z, frDirEnum dir);
-
-
-__host__ __device__
-void correct(frMIdx& x, frMIdx& y, frMIdx& z, frDirEnum& dir);
-
-__host__ __device__
-void correctU(frMIdx& x, frMIdx& y, frMIdx& z, frDirEnum& dir);
-
-__host__ __device__
-void reverse(frMIdx& x, frMIdx& y, frMIdx& z, frDirEnum& dir);
-
-
-// inline index function
-__host__ __device__
-inline double getCongCost(int demand, int supply)
+// For a node in a routed tree, there are at most 4 children nodes
+struct __align__(16) NodeStruct 
 {
-  return (demand * (4 / (1.0 + exp(supply - demand))) / (supply + 1));
-}
-
-// inline index function
-__host__ __device__
-inline frMIdx getIdx(frMIdx xIdx, frMIdx yIdx, frMIdx zIdx,
-              int xDim, int yDim, const bool* zDir)
-{
-  return (zDir[zIdx] == true) ? xIdx + yIdx * xDim + zIdx * xDim * yDim
-                              : yIdx + xIdx * yDim + zIdx * xDim * yDim;
-} 
-
-__host__ __device__
-inline frMIdx getIdx(frMIdx xIdx, frMIdx yIdx, int xDim, int yDim)
-{ 
-  return xIdx + yIdx * xDim;
-}
+  uint16_t nodeIdx; // 16-bit index
+  uint16_t parentIdx; // 16-bit index
+  uint16_t level; // 16-bit index
+  uint8_t childCnt;
+  uint8_t layerNum;
+  uint8_t minLayerNum;
+  uint8_t maxLayerNum;
+  uint16_t children[4];
+  uint32_t x;
+  uint32_t y;
+  int netId;
+};
 
 
-__host__ __device__
-inline bool isValid(frMIdx x, frMIdx y, frMIdx z, 
-             int xDim, int yDim, int zDim) 
-{
-  if (x < 0 || y < 0 || z < 0 || x >= xDim || y >= yDim || z >= zDim) {
-    return false;
-  }
+// Bit-level Operations
+__device__
+int getIdx__device(int x, int y, int z, int xDim, int yDim, int zDim);
+
+__device__
+bool getBit__device(uint64_t* d_cmap, unsigned idx, unsigned pos);
+
+__device__
+unsigned getBits__device(uint64_t* d_cmap, unsigned idx, unsigned pos, unsigned length);
+
+__device__
+void setBits__device(uint64_t* d_cmap, unsigned idx, unsigned pos, unsigned length, unsigned val);
+
+__device__
+void addToBits__device(uint64_t* d_cmap, unsigned idx, unsigned pos, unsigned length, unsigned val);
+
+
+// Note that we do not check if dir is aligned with the preferred direction
+__device__
+bool hasBlock__device(uint64_t* d_cmap, 
+  int xDim, int yDim, int zDim,
+  unsigned x, unsigned y, unsigned z, frDirEnum dir);
+
   
-  return true;
-}
+// Note that we do not check if dir is aligned with the preferred direction
+__device__
+unsigned getRawSupply__device(uint64_t* d_cmap, 
+  int xDim, int yDim, int zDim,
+  unsigned x, unsigned y, unsigned z, frDirEnum dir);
 
 
-// inline bit functions
-__host__ __device__
-inline bool getBit(const uint64_t* bits, frMIdx idx, frMIdx pos) {
-  return (bits[idx] >> pos) & 1;
-}
+// Note that we do not check if dir is aligned with the preferred direction
+__device__
+unsigned getRawDemand__device(uint64_t* d_cmap,
+  int xDim, int yDim, int zDim,
+  unsigned x, unsigned y, unsigned z, frDirEnum dir);
 
 
-__host__ __device__
-inline void setBit(uint64_t* bits, frMIdx idx, frMIdx pos) {
-  bits[idx] |= 1 << pos; 
-}
-
-__host__ __device__
-inline void resetBit(uint64_t* bits, frMIdx idx, frMIdx pos) {
-  bits[idx] &= ~(1 << pos); 
-}
-__host__ __device__
-inline unsigned getBits(const uint64_t* bits, frMIdx idx, frMIdx pos, unsigned length) {
-  auto tmp = bits[idx] & (((1ull << length) - 1) << pos);  // mask
-  return tmp >> pos;
-}
-
-__host__ __device__
-inline void setBits(uint64_t* bits, frMIdx idx, frMIdx pos, frUInt4 length, frUInt4 val) {
-  bits[idx] &= ~(((1ull << length) - 1) << pos);  // clear related bits to 0
-  bits[idx] |= ((uint64_t) val & ((1ull << length) - 1))
-                << pos;  // only get last length bits of val
-}
-
-__host__ __device__
-inline void addToBits(uint64_t* bits, frMIdx idx, frMIdx pos, frUInt4 length, frUInt4 val) {
-  auto tmp = getBits(bits, idx, pos, length) + val;
-  tmp = (tmp > (1u << length)) ? (1u << length) : tmp;
-  setBits(bits, idx, pos, length, tmp);
-}
-
-__host__ __device__
-inline void subToBits(uint64_t* bits, frMIdx idx, frMIdx pos, frUInt4 length, frUInt4 val)
-{
-  int tmp = (int) getBits(bits, idx, pos, length) - (int) val;
-  tmp = (tmp < 0) ? 0 : tmp;
-  setBits(bits, idx, pos, length, tmp);
-}
-
+__device__
+void addRawDemand__device(uint64_t* d_cmap,
+  int xDim, int yDim, int zDim,
+  unsigned x, unsigned y, unsigned z, frDirEnum dir, unsigned delta = 1);
 
 
 } // namespace drt
